+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to Pipelines will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Fixed
+- Remove PHP internal variables like $argv from the environment
+  variable maps in containers
+
 ## [0.0.19] - 2019-04-02
 ### Added
 - Suggestion to install the PHP YAML extension